services/agent_interface/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
jinja2
python-multipart
requests
